add api

+++ backend/app/routes/iot_routes.py
@@ -1,5 +1,7 @@
 from flask import Blueprint, request, jsonify
 from app.services import IoTService
+from app.utils import token_required
+
 
 iot_bp = Blueprint("iot", __name__)
 
@@ -13,8 +15,13 @@ def get_devices(room_id):
         return jsonify({"error": str(e)}), 404
 
 @iot_bp.route("/rooms/<room_id>/devices", methods=["POST"])
+@token_required
 def create_device(room_id):
     """Tạo thiết bị mới."""
+    role = request.role
+    if role != "admin":
+        return jsonify({"error": "Unauthorized"}), 403
+
     try:
         data = request.get_json()
         device = IoTService.create_device(room_id, data)
@@ -23,8 +30,13 @@ def create_device(room_id):
         return jsonify({"error": str(e)}), 400
 
 @iot_bp.route("/rooms/<room_id>/devices/<devices_name>", methods=["PUT"])
+@token_required
 def update_device(room_id, devices_name):
     """Cập nhật thiết bị."""
+    role = request.role
+    if role != "admin":
+        return jsonify({"error": "Unauthorized"}), 403
+    
     try:
         data = request.get_json()
         device = IoTService.update_device(room_id, devices_name, data)
@@ -33,8 +45,13 @@ def update_device(room_id, devices_name):
         return jsonify({"error": str(e)}), 400
 
 @iot_bp.route("/rooms/<room_id>/devices/<devices_name>", methods=["DELETE"])
+@token_required
 def delete_device(room_id, devices_name):
     """Xóa thiết bị."""
+    role = request.role
+    if role != "admin":
+        return jsonify({"error": "Unauthorized"}), 403
+    
     try:
         result = IoTService.delete_device(room_id, devices_name)
         return jsonify(result), 200
